* master:
  [cleanup] fix warnings for unused minInnerProduct
  [cleanup] fix warning with switch cases not in enum
  Revert "disable deprecation warnings in boost constants.hpp"

+++ dune/functions/functionspacebases/pqksubsampleddgbasis.hh
@@ -311,7 +311,7 @@ public:
     const auto& gridIndexSet = nodeFactory_->gridView().indexSet();
     const auto& element = node_->element();
 
-    switch (dim)
+    switch ((unsigned int)dim)
     {
       case 1:
       {
